Remove old migrations
