--- emails/ticket-close-warning.html
@@ -1,58 +1,0 @@
-<table align="center" border="0" cellpadding="0" cellspacing="0" width="100%" id="bodyTable">
-    <tr>
-        <td align="center" valign="top" id="bodyCell">
-            <!-- BEGIN TEMPLATE // -->
-            <table border="0" cellpadding="0" cellspacing="0" width="100%" class="templateContainer">
-                <tr>
-                    <td valign="top" id="templateBody">
-                        <table border="0" cellpadding="0" cellspacing="0" width="100%" class="mcnTextBlock" style="min-width:100%">
-                            <tbody class="mcnTextBlockOuter">
-                            <tr>
-                                <td valign="top" class="mcnTextBlockInner" style="padding-top:9px">
-                                    <table align="left" border="0" cellpadding="0" cellspacing="0" style="max-width:100%;min-width:100%" width="100%" class="mcnTextContentContainer">
-                                        <tbody>
-                                        <tr>
-
-                                            <td valign="top" class="mcnTextContent" style="padding-top:0;padding-right:18px;padding-bottom:9px;padding-left:18px">
-                                                <h1 style="text-align: center">Your request (#{%ticket_number%} {%ticket_subject%}) is awaiting response</h1>
-                                                <p>You are receiving this because your request for support is awaiting response but we have yet to hear back
-                                                    <br>
-                                                    If you still need to follow up on this ticket, please visit {%support_url%}
-                                                    <br>
-                                                    If this issue has been resolved, you may ignore this message and your ticket will be automatically closed within 24 hours.
-                                                </p>
-                                            </td>
-                                        </tr>
-                                        </tbody>
-                                    </table>
-                                    <table border="0" cellpadding="0" cellspacing="0" width="100%" class="mcnButtonBlock" style="min-width:100%">
-                                        <tbody class="mcnButtonBlockOuter">
-                                        <tr>
-                                            <td style="padding-top:0;padding-right:18px;padding-bottom:18px;padding-left:18px" valign="top" align="center" class="mcnButtonBlockInner">
-                                                <table border="0" cellpadding="0" cellspacing="0" class="mcnButtonContentContainer" style="border-collapse: separate !important;background-color: #26A6D6">
-                                                    <tbody>
-                                                    <tr>
-                                                        <td align="center" valign="middle" class="mcnButtonContent" style="font-family: Arial;font-size: 16px;padding: 15px">
-                                                            <a class="mcnButton " title="View Ticket" href="{%support_url%}" target="_blank" style="font-weight: bold;letter-spacing: 1px;line-height: 100%;text-align: center;text-decoration: none;color: #FFFFFF">View Open Tickets</a>
-                                                        </td>
-                                                    </tr>
-                                                    </tbody>
-                                                </table>
-                                            </td>
-                                        </tr>
-                                        </tbody>
-                                    </table>
-                                </td>
-                            </tr>
-                            </tbody>
-                        </table>
-                    </td>
-                </tr>
-                <tr>
-                    <td valign="top" id="templateFooter"></td>
-                </tr>
-            </table>
-            <!-- // END TEMPLATE -->
-        </td>
-    </tr>
-</table>
