+++ resources/emails/templates/ticket-closed.html
