+++ resources/emails/templates/ticket-created.html
